# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sidekeek API Service",
    "description": "Group of services provided to maintain sidekeek-client needs",
    "contact": "yourname@something.com",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/sideAPIkeek",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/confirmation": {
      "post": {
        "tags": ["RegisterController"],
        "summary": "confirm code",
        "description": "confirm code to validate email",
        "operationId": "confirmation",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "uniqueCode",
            "in" : "formData",
            "description": "unique_code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/confirmation"
            }
          }
        }
      }
    },
...
    "/sign-up": {
      "post": {
        "tags": ["RegisterController"],
        "summary": "sign up new seeker",
        "description": "sign up new user for seeker",
        "operationId": "sign-up",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "statTemp",
            "in" : "formData",
            "description": "random generated from client-side",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in" : "formData",
            "description": "email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in" : "formData",
            "description": "password",
            "required": true,
            "type": "string"
          },
          {
            "name": "confirmation",
            "in" : "formData",
            "description": "confirmation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/sign-up"
            }
          }
        }
      }
    },
...
    "/support": {
      "post": {
        "tags": ["UtilitiesController"],
        "summary": "support",
        "description": "support",
        "operationId": "support",
        "parameters": [
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "subject",
            "in" : "formData",
            "description": "subject",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in" : "formData",
            "description": "title",
            "required": true,
            "type": "string"
          },
          {
            "name": "message",
            "in" : "formData",
            "description": "message",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/post-simple"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "post-simple": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
...
    "sign-up": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        },
        "status": {
          "example": ""
        },
        "unique_code": {
          "example": ""
        },
        "email": {
          "example": ""
        }
      }
    },
...
    "confirmation": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        },
        "session": {
          "example": ""
        },
        "email": {
          "example": ""
        }
      }
    },
...
  }
}